requirements.txt
agora-token-builder==1.0.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
astroid==2.3.3
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
autopep8==1.6.0
bidict==0.23.1
certifi==2021.10.8
cffi==1.15.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==2.0.8
click==8.1.3
colorama==0.4.3
constantly==23.10.4
cryptography==44.0.2
cssselect2==0.6.0
daphne==4.1.2
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20230430
distlib==0.3.5
dj-database-url==1.0.0
Django==5.2
django-axes==6.1.0
django-cors-headers==4.7.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-import-export==3.3.7
django-jazzmin==2.6.0
django-login-required-middleware==0.4
django-ratelimit==4.1.0
django-widget-tweaks==1.4.5
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Faker==9.8.3
filelock==3.8.0
frozenlist==1.6.0
future==0.18.2
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.3
incremental==24.7.2
iniconfig==2.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.9.1
MarkupPy==1.14
mccabe==0.6.1
msgpack==1.1.0
multidict==6.4.3
mysql-connector-python==8.0.33
mysqlclient==2.1.0
Naked==0.1.31
num2words==0.5.12
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==25.0
Pillow==9.2.0
pipenv==2022.11.25
platformdirs==2.5.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
pydantic==2.11.4
pydantic_core==2.33.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.10.1
pylint==2.4.4
pyOpenSSL==25.0.0
PyPDF3==1.0.6
pytest==8.3.5
pytest-mock==3.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.2
python-engineio==4.12.0
python-social-auth==0.3.6
python-socketio==5.13.0
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.2.0
realtime==2.4.3
redis==5.2.1
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.1
service-identity==24.2.0
shellescape==3.8.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
storage3==0.11.3
StrEnum==0.4.15
stripe==4.0.2
supabase==2.15.1
supafunc==0.9.4
svglib==1.4.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.1.1
toml==0.10.2
tomli==2.2.1
tqdm==4.64.0
Twisted==24.11.0
txaio==23.1.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.7
virtualenv==20.16.3
virtualenv-clone==0.5.7
webencodings==0.5.1
websockets==14.2
whitenoise==6.9.0
wrapt==1.11.2
wsproto==1.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
yarl==1.20.0
zope.interface==7.2


gunicorn

